WIP: extracting cjs handler
